requirements.txt
Flask
python-docx
Werkzeug
